added a datehelper util
