+++ src/controllers/stock/stock.controller.js
